Created extension RSSFeed to refactor parsing of RSSFeed and remove the parsing codes from the EpisodesController.
