+++ Podcast/Podcast/Extensions/RSSFeed.swift
@@ -1,0 +1,29 @@
+//
+//  RSSFeed.swift
+//  Podcast
+//
+//  Created by Jaycell on 09/09/2019.
+//  Copyright © 2019 Jaycell. All rights reserved.
+//
+
+import FeedKit
+
+extension RSSFeed {
+    
+    func toEpisodes() -> [Episode] {
+        let imageUrl = iTunes?.iTunesImage?.attributes?.href
+        
+        var episodes = [Episode]() //blank Episode array
+        items?.forEach({ (feedItem) in
+            var episode = Episode(feedItem: feedItem)
+            
+            if episode.imageUrl == nil {
+                episode.imageUrl = imageUrl
+            }
+            
+            episodes.append(episode)
+        })
+        return episodes
+    }
+    
+}
